Player and Camera set up
Player and camera set up in order to have the player in the center of the screen and enough space around it to see the enemies.
The camera takes the height and distance values to calculate its position and the smoothness value to calculate at which speed it should follow the player.