Enemy behavior update
States have been assigned to the enemy.
They can now be idle, moving to player, move in circles and walking back.
The movement that is being tested now is the enemies moving towards the player and when inside of a certain radius they walk back. When they go past a further away radius they walk towards the player again.
This behavior loops infinitely so far.