StateMachine set up
I tried setting up a state machine to iterate around the various states that the enemy can have.
The idea was to have another script "EnemyStateHandler" that was assigned to each enemy and that took care of changing the state.
However, this was giving a lot of issue and Nullptr that I couldn't locate, therefore it has been implemented directly in the behavior of the enemies.
In this way it seems a big work around for a much simpler thing, but at least it's more tidy and readable.