Enemies in attack range
The state machine turned out to not be particularly effective.

Now the enemies in attacking range will start circulating the player and out of them, one will attack.
As soon as it's done attacking it will start moving back and a new random enemy will start attacking.
The other enemies that are not attacking or walking back are circulating around the player.