Enemies outside of camera viewport
The enemies outside of the camera viewport usually don't attack you, because the player is not able to see them.
Now if the enemies are outside of the screen they will not be able to attack the player.
For melee enemies this is not an issue because they just continue to do their typical behavior, but for ranged enemies this is very important.