Setup for different types of enemies
I added a new "MeleeEnemies" script so that it inherits from "EnemyBehavior".
In this way, there can be a variety of different types of enemies that all behave in a different way but that the EnemyManager doesn't have to care for each one of them.
The next step is to change the way the behaviors are handled in the EnemyManger.