Behavior clean up
The code has been cleaned up so that the EnemyManager doesn't need to do any calculation for the enemies but just tell it in which state it should be.
The EnemyBehavior has been updated so that it only calculats the common things to every enemy. The behaviors now depend on the type of enemy.